GrenadeExplosionRedo Part 1
Bloom and Emissive glow don't seem to work in our project but work by default in others. I've compared lighting settings & URP pipeline settings and found nothing that worked. Additional note the generate lighting in other projects takes like half a second but around 30 to over an hour in our build even though we only have a skybox for lighting which is a bit odd